moved types to separate files, added labels, removed unnecessary condition
