+++ src/components/TodoInfo/TodoInfo.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import cn from 'classnames';
 
-import { TodoWithUser } from '../../react-app-env';
+import { TodoWithUser } from '../../types/TodoWithUser';
 import { UserInfo } from '../UserInfo';
 
 type Props = {
