+++ src/components/TodoList/TodoList.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { TodoWithUser } from '../../react-app-env';
+import { TodoWithUser } from '../../types/TodoWithUser';
 import { TodoInfo } from '../TodoInfo';
 
 type Props = {
