+++ src/react-app-env.d.ts
@@ -1,19 +1,1 @@
 /// <reference types="react-scripts" />
-
-export interface User {
-  id: number;
-  name: string;
-  username: string;
-  email: string;
-}
-
-export interface Todo {
-  id: number;
-  title: string;
-  completed: boolean;
-  userId: number;
-}
-
-export interface TodoWithUser extends Todo {
-  user: User | null;
-}
