+++ src/types/Todo.ts
@@ -1,0 +1,6 @@
+export interface Todo {
+  id: number;
+  title: string;
+  completed: boolean;
+  userId: number;
+}
