+++ src/types/TodoWithUser.ts
@@ -1,0 +1,6 @@
+import { Todo } from "./Todo";
+import { User } from "./User";
+
+export interface TodoWithUser extends Todo {
+  user: User | null;
+}
